Circle Racing game in vb.net winform.
in this file you can be implemented circle racing game.
watch demo video which i provided in the files.
in here i design three circle and three lines in which circle is running .
once we click on start/play button all circle start running.once circle reach it's end point we decide which circle is reach end point first. after that we declare that circle as winner.

pretty simple and basic game for winform developer.